🧬 Migrate the tests about common functions to *inner_base* test module to avoid missing test.
